installer_dist/requirements.txt
psutil>=5.9.0
requests>=2.28.0
pystray>=0.19.0
Pillow>=9.0.0
plyer>=2.1.0
sqlite3
wmi>=1.5.1
configparser